safari-600.7-branch/LayoutTests/fast/selectors/long-adjacent-backtracking-expected.txt

Test very long backtracking of a direct adjacent chain.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("li.first+li+li+li+li+li+li+li+li+li+li+li+li+li+li+li+li+li+li+li+li~li").length is 20
PASS allItems.length is 41
PASS coloredCount is 20
PASS successfullyParsed is true

TEST COMPLETE
